Decrease frequency of dependabot updates
As we switched to nightly on DSharpPlus, the PRs are going to be spammed now with updates for this package. Decrease frequency of this checking from daily to weekly (defaults to Monday). This doesn't include security updates and forced checks, which will happen immediately. Also change labels and commit message to align more with how we manually label commits ourselves.